requirements.txt
openai
langchain
langchain-openai
langchain-pinecone
langchain_community
pinecone-client
python-dotenv
pypdf